+++ baybe/constraints/conditions.py
@@ -14,7 +14,6 @@ import numpy as np
 import pandas as pd
 from attrs import define, field
 from attrs.validators import in_, min_len
-from funcy import rpartial
 from numpy.typing import ArrayLike
 from typing_extensions import override
 
@@ -25,6 +24,7 @@ from baybe.serialization import (
     get_base_structure_hook,
     unstructure_base,
 )
+from baybe.utils.basic import to_tuple
 from baybe.utils.numerical import DTypeFloatNumpy
 
 if TYPE_CHECKING:
@@ -75,9 +75,9 @@ def _is_close(x: ArrayLike, y: ArrayLike, rtol: float, atol: float) -> np.ndarra
 _threshold_operators: dict[str, Callable] = {
     "<": ops.lt,
     "<=": ops.le,
-    "=": rpartial(_is_close, rtol=0.0),
-    "==": rpartial(_is_close, rtol=0.0),
-    "!=": rpartial(_is_not_close, rtol=0.0),
+    "=": partial(_is_close, rtol=0.0),
+    "==": partial(_is_close, rtol=0.0),
+    "!=": partial(_is_not_close, rtol=0.0),
     ">": ops.gt,
     ">=": ops.ge,
 }
@@ -166,10 +166,15 @@ class ThresholdCondition(Condition):
 
     def _make_operator_function(self):
         """Generate a function using operators to filter out undesired rows."""
-        func = rpartial(_threshold_operators[self.operator], self.threshold)
+
+        def evaluate(x: ArrayLike, /, **kwargs) -> Callable:
+            """Evaluate the condition on a given input."""
+            return _threshold_operators[self.operator](x, self.threshold, **kwargs)
+
         if self.operator in _valid_tolerance_operators:
-            func = rpartial(func, atol=self.tolerance)
-        return func
+            return partial(evaluate, atol=self.tolerance)
+
+        return evaluate
 
     @override
     def evaluate(self, data: pd.Series) -> pd.Series:
@@ -194,7 +199,7 @@ class SubSelectionCondition(Condition):
 
     # object variables
     _selection: tuple = field(
-        converter=tuple,
+        converter=to_tuple,
         # FIXME[typing]: https://github.com/python-attrs/attrs/issues/1197
         validator=[
             min_len(1),