+++ tests/insights/test_shap.py
@@ -7,15 +7,16 @@ from unittest import mock
 import numpy as np
 import pandas as pd
 import pytest
-from matplotlib import pyplot as plt
 from pytest import mark
 
 from baybe._optional.info import INSIGHTS_INSTALLED
-from baybe.exceptions import IncompatibleExplainerError
+from baybe.exceptions import IncompatibleExplainerError, NoMeasurementsError
 
 if not INSIGHTS_INSTALLED:
     pytest.skip("Optional insights package not installed.", allow_module_level=True)
 
+
+from matplotlib import pyplot as plt
 
 from baybe import insights
 from baybe._optional.insights import shap
@@ -174,7 +175,7 @@ def test_plots(ongoing_campaign: Campaign, use_comp_rep, plot_type):
 def test_updated_campaign_explanations(campaign, n_iterations, batch_size):
     """Test explanations for campaigns with updated measurements."""
     with pytest.raises(
-        ValueError,
+        NoMeasurementsError,
         match="The campaign does not contain any measurements.",
     ):
         SHAPInsight.from_campaign(campaign)